+++ src/server/http/Server.js
@@ -1,5 +1,6 @@
 import Request from './Request'
 import Response from './Response'
+import Headers from './Headers'
 import NoResponseError from '../../pipeline/NoResponseError'
 
 export default class Server {
@@ -39,13 +40,19 @@ export default class Server {
   }
 
   _parseRequest (nativeRequest) {
-    return new Request(nativeRequest.method, nativeRequest.url)
+    const headers = Object.keys(nativeRequest.headers)
+      .map((name) => [ name, nativeRequest.headers[name] ])
+      .reduce(
+        (headers, [name, value]) => headers.set(name, value),
+        new Headers()
+      )
+    return new Request(nativeRequest.method, nativeRequest.url, headers)
   }
 
   _writeResponse (response, nativeResponse) {
     response.headers.forEach(nativeResponse.setHeader.bind(nativeResponse))
     nativeResponse.writeHead(response.statusCode)
-    nativeResponse.end(response.body)
+    response.body.pipe(nativeResponse)
   }
 
   _errorResponse (error) {