+++ src/server/index.js
@@ -6,9 +6,12 @@ import Server from './http/Server'
 import Request from './http/Request'
 import Response from './http/Response'
 import Headers from './http/Headers'
+import ContentType from './http/ContentType'
+import ResponseBody from './http/ResponseBody'
 import serve from './http/serve'
 
 import * as errors from './http/errors'
+import * as middleware from './middleware'
 
 export {
   Pipeline,
@@ -19,7 +22,10 @@ export {
   Request,
   Response,
   Headers,
+  ContentType,
+  ResponseBody,
   serve,
 
-  errors
+  errors,
+  middleware
 }