+++ src/server/middleware/index.js
@@ -1,0 +1,11 @@
+import CompressionMiddleware from './CompressionMiddleware'
+import DeflateMiddleware from './DeflateMiddleware'
+import GzipMiddleware from './GzipMiddleware'
+import HttpEncoderMiddleware from './HttpEncoderMiddleware'
+
+export {
+  CompressionMiddleware,
+  DeflateMiddleware,
+  GzipMiddleware,
+  HttpEncoderMiddleware
+}